requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opencv-python==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
pydantic==2.9.2
sqlalchemy[asyncio]==2.0.34
asyncpg==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1